# openapi.yaml
openapi: 3.1.0
info:
  title: MinusPod API
  description: |
    REST API for managing podcast feeds with automatic ad detection and removal.


    This API allows you to:
    - Manage podcast feed subscriptions
    - View and manage processed episodes
    - Configure ad detection settings including AI model selection
    - Monitor system status and trigger cleanup operations
    - Manage cross-episode ad patterns with network and podcast scope
    - Submit corrections to improve ad detection accuracy
  version: 2.0.0
  contact:
    name: MinusPod
  license:
    name: MIT
servers:
  # Same-origin relative URL. Swagger UI served from /api/v1/docs uses
  # whatever host the browser is on, so "Try it out" always hits the
  # correct server: localhost in dev, the production domain in prod,
  # a Tailscale/Cloudflare URL on a tunnel, etc. No hardcoded host.
  - url: /api/v1
    description: Same-origin (current server)
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: '{baseUrl}/api/v1'
    description: Custom server
    variables:
      baseUrl:
        default: http://localhost:8000
        description: Base URL of the podcast server
...
paths:
...
  /feeds:
    get:
      tags:
        - Feeds
      summary: List all feeds
      description: Returns a list of all configured podcast feeds with metadata
      operationId: listFeeds
      responses:
        '200':
          description: List of feeds
          content:
            application/json:
              schema:
                type: object
                properties:
                  feeds:
                    type: array
                    items:
                      $ref: '#/components/schemas/Feed'
    post:
      tags:
        - Feeds
      summary: Add a new feed
      description: Add a new podcast feed by providing the source RSS URL
      operationId: addFeed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sourceUrl
              properties:
                sourceUrl:
                  type: string
                  format: uri
                  description: URL of the source RSS feed
                  example: https://feeds.example.com/podcast.xml
                slug:
                  type: string
                  description: Optional custom slug (auto-generated if not provided)
                  example: my-podcast
                maxEpisodes:
                  type: integer
                  minimum: 10
                  maximum: 500
                  description: Maximum number of episodes served in the RSS feed (default 300)
      responses:
        '201':
          description: Feed created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug:
                    type: string
                  sourceUrl:
                    type: string
                  feedUrl:
                    type: string
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Feed with slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /feeds/import-opml:
    post:
      tags:
        - Feeds
      summary: Import feeds from OPML file
      description: |
        Import podcast feeds from an OPML file upload. Accepts .opml or .xml files.
        Automatically generates slugs from feed titles and refreshes the first 5 imported feeds.
      operationId: importOpml
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - opml
              properties:
                opml:
                  type: string
                  format: binary
                  description: OPML file (.opml or .xml)
      responses:
        '201':
          description: Feeds imported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                    description: Number of feeds successfully imported
                  skipped:
                    type: integer
                    description: Number of feeds skipped (already exist)
                  failed:
                    type: integer
                    description: Number of feeds that failed to import
                  feeds:
                    type: object
                    properties:
                      imported:
                        type: array
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                            slug:
                              type: string
                      skipped:
                        type: array
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                            slug:
                              type: string
                            reason:
                              type: string
                      failed:
                        type: array
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                            error:
                              type: string
        '200':
          description: No feeds were imported (all skipped or failed)
        '400':
          description: Invalid OPML file or no file provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /feeds/export-opml:
    get:
      tags:
        - Feeds
      summary: Export feeds as OPML file
      description: |
        Export all subscribed podcast feeds as an OPML XML file.
        Use mode=original for upstream source URLs (backup/migration).
        Use mode=modified for MinusPod-served ad-free feed URLs (import into podcast apps).
      operationId: exportOpml
      parameters:
        - name: mode
          in: query
          description: 'URL mode: "original" for upstream source URLs, "modified" for MinusPod ad-free feed URLs'
          schema:
            type: string
            enum: [original, modified]
            default: original
      responses:
        '200':
          description: OPML file download
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 'Attachment filename, e.g. attachment; filename="minuspod-feeds.opml"'
          content:
            application/xml:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid mode parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /feeds/{slug}:
    get:
      tags:
        - Feeds
      summary: Get a single feed
      description: Returns details for a specific podcast feed
      operationId: getFeed
      parameters:
        - $ref: '#/components/parameters/slug'
      responses:
        '200':
          description: Feed details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Feeds
      summary: Delete a feed
      description: Delete a podcast feed and all associated data
      operationId: deleteFeed
      parameters:
        - $ref: '#/components/parameters/slug'
      responses:
        '200':
          description: Feed deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  slug:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Feeds
      summary: Update feed settings
      description: Update podcast feed settings like network override, DAI platform, and processing overrides
      operationId: updateFeed
      parameters:
        - $ref: '#/components/parameters/slug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                networkIdOverride:
                  type:
                    - string
                    - 'null'
                  description: Override network ID (null to use auto-detected)
                daiPlatform:
                  type: string
                  description: DAI platform identifier (e.g., megaphone, acast)
                autoProcessOverride:
                  type:
                    - boolean
                    - 'null'
                  description: Override auto-process setting (null uses global)
                maxEpisodes:
                  type: integer
                  minimum: 10
                  maximum: 500
                  description: Maximum number of episodes served in the RSS feed (default 300)
      responses:
        '200':
          description: Feed updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /feeds/{slug}/refresh:
    post:
      tags:
        - Feeds
      summary: Refresh a single feed
      description: |
        Fetch the latest RSS feed and update the episode list. By default
        this sends a conditional GET -- if upstream returns 304 Not
        Modified, the refresh is a no-op. Pass `{"force": true}` in the
        body to bypass the stored ETag / Last-Modified headers and do a
        full fetch, useful when an upstream CDN is falsely returning 304
        against its own new content.
      operationId: refreshFeed
      parameters:
        - $ref: '#/components/parameters/slug'
      requestBody:
        required: false
        description: Optional force-refresh flag. Omit for default conditional GET.
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  description: |
                    When true, clears the podcast's stored ETag +
                    Last-Modified values before fetching so upstream
                    cannot short-circuit with 304 Not Modified.
                  default: false
              additionalProperties: false
      responses:
        '200':
          description: Feed refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug:
                    type: string
                  message:
                    type: string
                  episodeCount:
                    type: integer
                  lastRefreshed:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
...
  /feeds/{slug}/artwork:
    get:
      tags:
        - Feeds
      summary: Get feed artwork
      description: Returns the cached podcast artwork image
      operationId: getFeedArtwork
      parameters:
        - $ref: '#/components/parameters/slug'
      responses:
        '200':
          description: Artwork image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Flask session cookie issued by ``POST /auth/login``. Every mutating
        endpoint additionally requires a matching ``X-CSRF-Token`` header
        (double-submit). Public endpoints (``/auth/login``, ``/auth/logout``,
        ``/auth/status``, ``/health``, ``/health/live``) declare ``security: []``.
  parameters:
    slug:
      name: slug
      in: path
      required: true
      description: Podcast feed slug identifier
      schema:
        type: string
        example: my-podcast
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...